requirements.txt
<<<<<<< HEAD
google-generativeai==0.3.2
speechrecognition==3.10.0
pyttsx3==2.90
flask==2.3.3
python-dotenv==1.0.0
pyaudio==0.2.14
supabase==2.5.0
requests==2.31.0
python-jose==3.3.0
flask-session==0.5.0
=======
google-generativeai==0.3.2
speechrecognition==3.10.0
pyttsx3==2.90
flask==2.3.3
python-dotenv==1.0.0
pyaudio==0.2.14
>>>>>>> 44168a780ffe20832cc5b64a57dfbe35a56e354d
